+++ docs/examples/SDropdown/compact.vue
@@ -2,27 +2,19 @@
 import { usePreview } from '~/composables/usePreview';
 
 const { controls } = usePreview({
-    mode: 'playground',
-    component: 'SDropdown',
     props: {
         variant: {
             type: 'select',
             options: ['default', 'compact'],
-            default: 'default',
+            default: 'compact',
             label: 'variant',
-        },
-        placement: {
-            type: 'select',
-            options: ['bottom-start', 'bottom', 'bottom-end', 'top-start', 'top', 'top-end'],
-            default: 'bottom-start',
-            label: 'placement',
         },
     },
 });
 </script>
 
 <template>
-    <SDropdown :variant="controls.variant" :placement="controls.placement">
+    <SDropdown :variant="controls.variant">
         <template #reference>Options</template>
 
         <SDropdownItem>Edit</SDropdownItem>